+++ deepinv/datasets/lidc_idri.py
@@ -29,6 +29,8 @@ except:
 class LidcIdriSliceDataset(torch.utils.data.Dataset):
     """Dataset for `LIDC-IDRI <https://www.cancerimagingarchive.net/collection/lidc-idri/>`_ that provides access to CT image slices.
 
+    Published in :footcite:ps:`armato2011lung`.
+
     | The Lung Image Database Consortium image collection (LIDC-IDRI) consists
     | of diagnostic and lung cancer screening thoracic computed tomography (CT)
     | scans with marked-up annotated lesions.
@@ -71,6 +73,10 @@ class LidcIdriSliceDataset(torch.utils.data.Dataset):
             batch = next(iter(dataloader))
             print(batch.shape)
 
+
+    :References:
+
+    .. footbibliography::
     """
 
     class SliceSampleIdentifier(NamedTuple):