+++ deepinv/models/dsccp.py
@@ -9,7 +9,7 @@ class DScCP(Denoiser):
     DScCP denoiser network.
 
     The network architecture is based on the paper
-    `Unfolded proximal neural networks for robust image Gaussian denoising <https://arxiv.org/abs/2308.03139>`_,
+    `Unfolded proximal neural networks for robust image Gaussian denoising :footcite:ps:`le2024unfolded`.
     and has an unrolled architecture based on the fast Chambolle-Pock algorithm using strong convexity.
     DScCP stands for Deep Strongly Convex Chambolle Pock.
 
@@ -20,6 +20,11 @@ class DScCP(Denoiser):
     :param int n_channels_per_layer: number of channels per convolutional layer.
     :param str pretrained: 'download' to download pretrained weights, or path to local weights file.
     :param torch.device, str device: 'cuda' or 'cpu'.
+
+    |sep|
+
+    :References:
+     .. footbibliography::
     """
 
     def __init__(