+++ deepinv/sampling/langevin.py
@@ -238,8 +238,7 @@ class ULA(MonteCarlo):
     r"""
     Projected Plug-and-Play Unadjusted Langevin Algorithm.
 
-    The algorithm runs the following markov chain iteration
-    (Algorithm 2 from https://arxiv.org/abs/2103.04715):
+    The algorithm runs the following markov chain iteration (Algorithm 2 from :footcite:ps:`laumont2022bayesian`):
 
     .. math::
 
@@ -277,6 +276,11 @@ class ULA(MonteCarlo):
         of the function g_statistic. By default, it is the identity function (lambda x: x),
         and thus the sampler computes the posterior mean and variance.
     :param bool verbose: prints progress of the algorithm.
+
+    |sep|
+
+    :References:
+        .. footbibliography::
 
     """
 
@@ -371,8 +375,7 @@ class SKRock(MonteCarlo):
     Obtains samples of the posterior distribution using an orthogonal Runge-Kutta-Chebyshev stochastic
     approximation to accelerate the standard Unadjusted Langevin Algorithm.
 
-    The algorithm was introduced in "Accelerating proximal Markov chain Monte Carlo by using an explicit stabilised method"
-    by L. Vargas, M. Pereyra and K. Zygalakis (https://arxiv.org/abs/1908.08845)
+    The algorithm was introduced in :footcite:ps:`pereyra2020accelerating`.
 
     - SKROCK assumes that the denoiser is :math:`L`-Lipschitz differentiable
     - For convergence, SKROCK required step_size smaller than :math:`\frac{1}{L+\|A\|_2^2}`
@@ -397,6 +400,11 @@ class SKRock(MonteCarlo):
     :param Callable g_statistic: The sampler will compute the posterior mean and variance
         of the function g_statistic. By default, it is the identity function (lambda x: x),
         and thus the sampler computes the posterior mean and variance.
+
+    |sep|
+
+    :References:
+        .. footbibliography::
 
     """