+++ docs/source/conf.py
@@ -44,8 +44,13 @@ extensions = [
     "sphinx_copybutton",
     "sphinx_design",
     "sphinx_sitemap",
+    "sphinxcontrib.bibtex",
 ]
+
+bibtex_bibfiles = ["refs.bib"]
+bibtex_default_style = "plain"
 copybutton_exclude = ".linenos, .gp"
+bibtex_tooltips = True
 
 intersphinx_mapping = {
     "numpy": ("https://numpy.org/doc/stable/", None),
@@ -64,6 +69,8 @@ autodoc_preserve_defaults = True
 nitpicky = True
 # Create link to the API in the auto examples
 autodoc_inherit_docstrings = False
+# For bibtex
+bibtex_footbibliography_backrefs = True
 # for sitemap
 html_baseurl = "https://deepinv.github.io/deepinv/"
 # the default scheme makes for wrong urls so we specify it properly here
@@ -270,3 +277,8 @@ rst_prolog = """
 
    <hr />
 """
+
+napoleon_custom_sections = [
+    ("Reference", "params_style"),  # Sphinx ≥ 3.5
+    # ("Reference", "Parameters"),   # fallback syntax for very old Sphinx (<3.5)
+]